revert StartAudio hook back to original implementation and move room audio observable into shared room.ts
