+++ packages/core/src/components/startAudio.ts
@@ -1,22 +1,7 @@
 import type { Room } from 'livekit-client';
-import { RoomEvent } from 'livekit-client';
-import { map } from 'rxjs';
 import log from '../logger';
-import { observeRoomEvents } from '../observables/room';
+import { roomAudioPlaybackAllowedObservable } from '../observables/room';
 import { prefixClass } from '../styles-interface';
-
-/**
- * Returns whether or not audio playback is allowed in the current context.
- * @internal
- */
-export function roomAudioPlaybackAllowedObservable(room: Room) {
-  const observable = observeRoomEvents(room, RoomEvent.AudioPlaybackStatusChanged).pipe(
-    map((room) => {
-      return { canPlayAudio: room.canPlaybackAudio };
-    }),
-  );
-  return observable;
-}
 
 export function setupStartAudio() {
   const handleStartAudioPlayback = async (room: Room) => {
