+++ packages/react/src/components/controls/StartAudio.tsx
@@ -5,6 +5,12 @@ import { useRoomContext } from '../../context';
 import { useObservableState } from '../../hooks/internal/useObservableState';
 import { mergeProps } from '../../utils';
 
+/** @alpha */
+export interface UseStartAudioProps {
+  room: Room;
+  props: React.ButtonHTMLAttributes<HTMLButtonElement>;
+}
+
 /**
  * In many browsers to start audio playback, the user must perform a user-initiated event such as clicking a button.
  * The `useStatAudio` hook returns an object with a boolean `canPlayAudio` flag
@@ -14,23 +20,30 @@ import { mergeProps } from '../../utils';
  * @see Autoplay policy on MDN web docs for more info: {@link https://developer.mozilla.org/en-US/docs/Web/API/Web_Audio_API/Best_practices#autoplay_policy}
  * @alpha
  */
-export function useStartAudio(room: Room) {
-  const { roomAudioPlaybackAllowedObservable, handleStartAudioPlayback } = React.useMemo(
+export function useStartAudio({ room, props }: UseStartAudioProps) {
+  const { className, roomAudioPlaybackAllowedObservable, handleStartAudioPlayback } = React.useMemo(
     () => setupStartAudio(),
     [],
   );
-
-  function startAudio() {
-    handleStartAudioPlayback(room);
-  }
-
   const observable = React.useMemo(
     () => roomAudioPlaybackAllowedObservable(room),
     [room, roomAudioPlaybackAllowedObservable],
   );
   const { canPlayAudio } = useObservableState(observable, { canPlayAudio: false });
 
-  return { canPlayAudio, startAudio };
+  const mergedProps = React.useMemo(
+    () =>
+      mergeProps(props, {
+        className,
+        onClick: () => {
+          handleStartAudioPlayback(room);
+        },
+        style: { display: canPlayAudio ? 'none' : 'block' },
+      }),
+    [props, className, canPlayAudio, handleStartAudioPlayback, room],
+  );
+
+  return { mergedProps, canPlayAudio };
 }
 
 /** @public */
@@ -55,20 +68,7 @@ export interface AllowAudioPlaybackProps extends React.ButtonHTMLAttributes<HTML
  */
 export function StartAudio({ label = 'Allow Audio', ...props }: AllowAudioPlaybackProps) {
   const room = useRoomContext();
-  const { canPlayAudio, startAudio } = useStartAudio(room);
-  const { className } = React.useMemo(() => setupStartAudio(), []);
-
-  const mergedProps = React.useMemo(
-    () =>
-      mergeProps(props, {
-        className,
-        onClick: () => {
-          startAudio();
-        },
-        style: { display: canPlayAudio ? 'none' : 'block' },
-      }),
-    [props, className, canPlayAudio, startAudio],
-  );
+  const { mergedProps } = useStartAudio({ room, props });
 
   return <button {...mergedProps}>{label}</button>;
 }
